+++ cypress/e2e/homepage.cy.ts
@@ -1,6 +1,9 @@
 describe("Homepage Products - problem_user", () => {
   beforeEach(() => {
-    cy.visit("/");
+    cy.fixture("usuarios").then((dados) => {
+      usuarios = dados;
+      cy.visit("/");
+    });
   });
 
   it("Deve exibir imagens iguais para cada produto", () => {
@@ -11,6 +14,7 @@ describe("Homepage Products - problem_user", () => {
   it("Tentativa de adicionar qualquer produto ao carrinho", () => {
     cy.loginSuccessfully(usuarios.problem_user, usuarios.password);
     cy.addProductToCart();
+
   });
 
   it('Produto 4 exibe "Item Not Found" e valor com caracteres inválidos');